Move tests to integration tests
