+++ integration-test/src/jvmTest/kotlin/org/spekframework/spek2/FeatureBodyAliasesTest.kt
@@ -1,0 +1,34 @@
+package org.spekframework.spek2
+
+import org.spekframework.spek2.style.gherkin.FeatureBody
+import kotlin.test.assertSame
+
+object FeatureBodyAliasesTest : Spek({
+
+    val groupBody by memoized { FakeGroupBody() }
+    val featureBody by memoized { FeatureBody(groupBody) }
+
+    test("beforeEachScenario should call beforeEachGroup") {
+        val fixture = { println("hello") }
+        featureBody.beforeEachScenario(fixture)
+        assertSame(fixture, groupBody.beforeEachGroup)
+    }
+
+    test("afterEachScenario should call AfterEachGroup") {
+        val fixture = { println("hello") }
+        featureBody.afterEachScenario(fixture)
+        assertSame(fixture, groupBody.afterEachGroup)
+    }
+
+    test("beforeFeature should call BeforeGroup") {
+        val fixture = { println("hello") }
+        featureBody.beforeFeature(fixture)
+        assertSame(fixture, groupBody.beforeGroup)
+    }
+
+    test("afterFeature should call AfterGroup") {
+        val fixture = { println("hello") }
+        featureBody.afterFeature(fixture)
+        assertSame(fixture, groupBody.afterGroup)
+    }
+})
