+++ spek-dsl/build.gradle.kts
@@ -76,20 +76,6 @@ kotlin {
                 implementation(kotlin("stdlib-common"))
             }
         }
-
-        commonTest {
-            dependencies {
-                api(kotlin("test-common"))
-                api(kotlin("test-annotations-common"))
-            }
-        }
-
-        val jvmTest by existing {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation(kotlin("test-junit"))
-            }
-        }
         
         jvm {
             compilations["main"].defaultSourceSet {
